@@ -1,6 +1,6 @@
 module github.com/vitaliy-ukiru/fsm-telebot
 
-go 1.17
+go 1.21
 
 require (
 	github.com/stretchr/testify v1.8.1
