@@ -6,12 +6,12 @@ import (
 
 // Filter object. Needs for graceful works with state filters.
 type Filter struct {
-	Endpoint interface{}
+	Endpoint any
 	States   []State
 }
 
 // F returns new Filter object.
-func F(endpoint interface{}, states ...State) Filter {
+func F(endpoint any, states ...State) Filter {
 	if len(states) == 0 {
 		states = []State{DefaultState}
 	}
@@ -20,7 +20,16 @@ func F(endpoint interface{}, states ...State) Filter {
 
 // TelebotHandlerForState creates tele.Handler with local filter for given state.
 func (m *Manager) TelebotHandlerForState(want State, handler Handler) tele.HandlerFunc {
-	return m.TelebotHandlerForStates(handler, want)
+	return m.HandlerAdapter(func(c tele.Context, state Context) error {
+		s, err := state.State()
+		if err != nil {
+			return &ErrHandlerState{Handler: "Manager.ForState", Err: err}
+		}
+		if Is(s, want) {
+			return handler(c, state)
+		}
+		return nil
+	})
 }
 
 // TelebotHandlerForStates creates a handler with local filter
@@ -39,7 +48,11 @@ func (m *Manager) TelebotHandlerForStates(h Handler, states ...State) tele.Handl
 }
 
 func (f Filter) CallbackUnique() string {
-	switch end := f.Endpoint.(type) {
+	return getEndpoint(f.Endpoint)
+}
+
+func getEndpoint(e any) string {
+	switch end := e.(type) {
 	case string:
 		return end
 	case tele.CallbackEndpoint:
