@@ -40,7 +40,7 @@ const prettyJson = "pretty_json"
 
 func (j PrettyJson) ProviderName() string { return prettyJson }
 
-func (j PrettyJson) Encode(v interface{}) ([]byte, error) {
+func (j PrettyJson) Encode(v any) ([]byte, error) {
 	buff := new(bytes.Buffer)
 
 	e := json.NewEncoder(buff)
@@ -54,7 +54,7 @@ func (j PrettyJson) Encode(v interface{}) ([]byte, error) {
 	return buff.Bytes(), nil
 }
 
-func (j PrettyJson) Decode(data []byte, v interface{}) error {
+func (j PrettyJson) Decode(data []byte, v any) error {
 	buff := bytes.NewReader(data)
 
 	d := json.NewDecoder(buff)
@@ -114,7 +114,7 @@ type record struct {
 	Data  map[string]json.RawMessage `json:"data"`
 }
 
-func (j PrettyJson) tryDecodeB64(enc *b64.Encoding, src []byte) ([]byte, bool) {
+func (PrettyJson) tryDecodeB64(enc *b64.Encoding, src []byte) ([]byte, bool) {
 	if src[0] != '"' && src[len(src)-1] != '"' {
 		return nil, false
 	}
@@ -128,7 +128,7 @@ func (j PrettyJson) tryDecodeB64(enc *b64.Encoding, src []byte) ([]byte, bool) {
 	return buf[:n], true
 }
 
-func (j PrettyJson) convertTo(storage file.ChatsStorage) jsonStorage {
+func (PrettyJson) convertTo(storage file.ChatsStorage) jsonStorage {
 	result := make(jsonStorage)
 	for chatId, usersStorage := range storage {
 		usersData := make(map[int64]record)
